requirements.txt
arviz==0.18.0
pymc
typer
rpy2
biopython